% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescribeR.R
\name{data_describer}
\alias{data_describer}
\title{data_describer function}
\usage{
data_describer(data)
}
\arguments{
\item{data}{A dataframe you want to describe}
}
\value{
A dataframe
}
\description{
data_describer function
}
\examples{
data_describer(tidyr::billboard)
}
