% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDescribeR.R
\name{data_describer}
\alias{data_describer}
\title{data_describer function}
\usage{
data_describer(data, example_number)
}
\arguments{
\item{data}{A dataframe you want to describe}

\item{example_number}{The amount of example values you want to include in the output, must be a numeric input}
}
\value{
A dataframe
}
\description{
data_describer function
}
\examples{
data_describer(data = tidyr::billboard, example_number = 3)
}
